FAC-814: Change type of questionWeight to int

+++ flickit-assessment-data/src/main/java/org/flickit/assessment/data/jpa/kit/attribute/AttributeQuestionView.java
@@ -7,7 +7,7 @@ public interface AttributeQuestionView {
 
     long getQuestionId();
 
-    Double getQuestionWeight();
+    int getQuestionWeight();
 
     AnswerJpaEntity getAnswer();
